Backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
django-filter==23.5
